made Vector2D support Position.

+++ CodenameG/src/edu/chl/codenameg/model/Vector2D.java
@@ -1,23 +1,23 @@
 package edu.chl.codenameg.model;
 
 public class Vector2D {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Vector2D(int x,int y){
+	public Vector2D(double x,double y){
 		this.setX(x);
 		this.setY(y);
 	}
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	public void add(Vector2D v2d){
